implement minmaxing for move choice
There were a couple of major changes made here. The first is this:
before, moves were scored based on the quality of the move. The
heuristic would score moves based on how good the move was
compared to its surrounding board. Now, moves aren't scored, but
the entire board is; this allows the minimax algorithm to play
properly in the test case provided and is also probably (?) more
consistent.

Another change that was an effect of the first one is that now the
helper methods to help score moves have been moved to board.cpp, because
we're scoring boards instead of moves.

The biggest change, however, is that now Kurellameme will actually use
minmaxing. Basically, it will look at all the moves it can make and then
look at all the moves the opponent will make, and will choose to play
the line that minimizes the damage if the opponent plays optimally
(according to our heuristic, which, as of right now, is definitely NOT
a safe thing to bet on).